ADOP-647 changes
