+++ src/main/steps/home/get.ts
@@ -11,7 +11,7 @@ import {
   CHECK_ANSWERS_URL,
   CONFIRM_JOINT_APPLICATION,
   HUB_PAGE,
-  PAY_YOUR_FEE,
+  // PAY_YOUR_FEE,
   SENT_TO_APPLICANT2_FOR_REVIEW,
   TASK_LIST_URL,
 } from '../urls';
@@ -41,9 +41,9 @@ const applicant1RedirectPageSwitch = (caseState: State, userCase: Partial<Case>,
     case State.Submitted: {
       return APPLICATION_SUBMITTED;
     }
-    case State.AwaitingPayment: {
-      return PAY_YOUR_FEE;
-    }
+    // case State.AwaitingPayment: {
+    //   return PAY_YOUR_FEE;
+    // }
     case State.AwaitingAos:
     case State.AwaitingConditionalOrder:
     case State.AosDrafted:
