+++ src/main/steps/review-pay-submit/reviewPaySubmitSequence.test.ts
@@ -2,7 +2,7 @@ import { reviewPaySubmitSequence } from './reviewPaySubmitSequence';
 
 describe('reviewPaySubmitSequence', () => {
   test('should contain 4 entries in review pay and submit screen sequence', () => {
-    expect(reviewPaySubmitSequence).toHaveLength(5);
+    expect(reviewPaySubmitSequence).toHaveLength(4);
 
     expect(reviewPaySubmitSequence[0].url).toBe('/review-pay-submit/equality');
     expect(reviewPaySubmitSequence[0].showInSection).toBe('reviewPaySubmit');
@@ -14,14 +14,14 @@ describe('reviewPaySubmitSequence', () => {
 
     expect(reviewPaySubmitSequence[2].url).toBe('/review-pay-submit/statement-of-truth');
     expect(reviewPaySubmitSequence[2].showInSection).toBe('reviewPaySubmit');
-    expect(reviewPaySubmitSequence[2].getNextStep({})).toBe('/review-pay-submit/payment/pay-your-fee');
+    expect(reviewPaySubmitSequence[2].getNextStep({})).toBe('/review-pay-submit/payment/payment-callback');
 
-    expect(reviewPaySubmitSequence[3].url).toBe('/review-pay-submit/payment/pay-your-fee');
+    expect(reviewPaySubmitSequence[3].url).toBe('/review-pay-submit/payment/payment-callback');
     expect(reviewPaySubmitSequence[3].showInSection).toBe('reviewPaySubmit');
-    expect(reviewPaySubmitSequence[3].getNextStep({})).toBe('/review-pay-submit/payment/payment-callback');
+    expect(reviewPaySubmitSequence[3].getNextStep({})).toBe('/task-list');
 
-    expect(reviewPaySubmitSequence[4].url).toBe('/review-pay-submit/payment/payment-callback');
-    expect(reviewPaySubmitSequence[4].showInSection).toBe('reviewPaySubmit');
-    expect(reviewPaySubmitSequence[4].getNextStep({})).toBe('/task-list');
+    // expect(reviewPaySubmitSequence[4].url).toBe('/review-pay-submit/payment/payment-callback');
+    // expect(reviewPaySubmitSequence[4].showInSection).toBe('reviewPaySubmit');
+    // expect(reviewPaySubmitSequence[4].getNextStep({})).toBe('/task-list');
   });
 });
